pyproject.toml
[tool.poetry]
name = "system-usage-monitor"
version = "0.1.0"
description = "A backend service for tracking and saving system performance metrics using FastAPI and InfluxDB."
authors = ["Murat Mert Güler <muratmertguler@hotmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
influxdb-client = "^1.47.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
